cant delete token if dolversion < 23

+++ class/providers/AbstractPDPProvider.class.php
@@ -429,6 +429,15 @@ abstract class AbstractPDPProvider
 
 		// Build service name depending on environment
 		$serviceName = $this->config['dol_prefix'] . '_' . ($this->config['live'] ? 'PROD' : 'TEST');
+		// For backward compatibility with Dolibarr versions < 23.0.0
+
+		if (version_compare(DOL_VERSION, '23.0.0', '<')) {
+			require_once DOL_DOCUMENT_ROOT."/core/lib/admin.lib.php";
+			dolibarr_del_const($this->db, $serviceName.'_TOKEN', $conf->entity);
+			dolibarr_del_const($this->db, $serviceName.'_REFRESH', $conf->entity);
+			dolibarr_del_const($this->db, $serviceName.'_EXPIRE', $conf->entity);
+			return true;
+		}
 
 		// Check if a token already exists for this service
 		$sql_check = "DELETE FROM ".MAIN_DB_PREFIX."oauth_token
